requirements.txt
PyYAML
logzero
pandas
numpy
scikit-learn
joblib
matplotlib
jupyter